adding usage + refactoring devtc client

+++ devtc_client/devtcApp.py
@@ -1,6 +1,7 @@
 import os
 import shutil
 import stat
+import sys
 import tarfile
 import zipfile
 
@@ -12,27 +13,34 @@ DEVTC_BIN_PATH = "{}/bin".format(DEVTC_HOME_PATH)
 DEVTC_ENV_VARIABLE = "DEVTC_HOME"
 
 
-def set_env_variable():
-    print("Setting devtc environment variable {}={}...".format(DEVTC_ENV_VARIABLE, DEVTC_HOME_PATH))
+def modify_env_variables(install=True):
+    option = "Setting" if install else "Resetting"
+    print("{} devtc environment variable {}={}...".format(option, DEVTC_ENV_VARIABLE, DEVTC_HOME_PATH))
+
+    bash_rc_path = "{}/.bashrc".format(HOME_PATH)
+    bash_rc_backup_path = "{}.bak".format(bash_rc_path)
+    env_variable_content = '\n' + 'export {}={}\nexport PATH=$PATH:${}/bin\n'.format(DEVTC_ENV_VARIABLE, DEVTC_HOME_PATH, DEVTC_ENV_VARIABLE)
 
     try:
-        bash_rc_path = "{}/.bashrc".format(HOME_PATH)
+
         file_exists = os.path.isfile(bash_rc_path)
         if file_exists:
-            bash_rc_tmp_path = "{}_tmp".format(bash_rc_path)
+            shutil.copyfile(bash_rc_path, bash_rc_backup_path)
 
-            with open(bash_rc_path, 'rt') as bash_rc_file:
+            with open(bash_rc_path, 'r') as bash_rc_file:
                 bash_rc_content = bash_rc_file.read()
-                if DEVTC_ENV_VARIABLE not in bash_rc_content:
-                    new_bash_rc_content = bash_rc_content + '\n' + 'export {}={}\nexport PATH=$PATH:${}/bin\n'.format(DEVTC_ENV_VARIABLE, DEVTC_HOME_PATH, DEVTC_ENV_VARIABLE)
-                    with open(bash_rc_tmp_path, 'wt') as bash_rc_tmp_file:
-                        bash_rc_tmp_file.write(new_bash_rc_content)
-                        os.rename(bash_rc_tmp_path, bash_rc_path)
 
-            print("devtc environment set successfully")
+            with open(bash_rc_path, 'wt') as bash_rc_file:
+                set_env_var = install and DEVTC_ENV_VARIABLE not in bash_rc_content
+                new_bash_rc_content = bash_rc_content + env_variable_content if set_env_var else bash_rc_content.replace(env_variable_content, '')
+                bash_rc_file.write(new_bash_rc_content)
+
+                print("devtc environment successfully modified")
         else:
             print("Could not set devtc environment variable - missing .bashrc user configuration")
     except RuntimeError as e:
+        if os.path.isfile(bash_rc_backup_path):
+            os.rename(bash_rc_backup_path, bash_rc_path)
         print("Could not set devtc environment variable", e)
 
 
@@ -41,7 +49,7 @@ def install_devtc():
         if HOME_PATH is None:
             raise ValueError("Could not install devtc - as no HOME Variable is set")
 
-        if os.path.isdir(DEVTC_HOME_PATH):
+        if is_devtc_installed():
             print("devtc already installed in {}".format(DEVTC_HOME_PATH))
             return
 
@@ -50,7 +58,7 @@ def install_devtc():
         os.mkdir(DEVTC_HOME_PATH)
         os.mkdir(DEVTC_BIN_PATH)
 
-        set_env_variable()
+        modify_env_variables()
 
         print("devtc installed successfully")
     except RuntimeError as e:
@@ -58,13 +66,34 @@ def install_devtc():
 
 
 def remove_devtc():
+    if not is_devtc_installed():
+        print("devtc is already removed!")
+        return
+
     shutil.rmtree(DEVTC_HOME_PATH, ignore_errors=True)
-    # todo unset environment variable
-    print("devtc removed successfully")
+    modify_env_variables(False)
+    print("devtc successfully removed")
 
 
-def install_tool(name, release_version, url, package_extension, package_binary_path):
+def is_devtc_installed():
+    return os.path.isdir(DEVTC_HOME_PATH)
+
+
+def process_devtc(process):
+    if process is None or process not in ["install", "remove"]:
+        print("Did you mean install?")
+        usage()
+        return
+
+    install_devtc() if process == "install" else remove_devtc()
+
+
+def install_tool(name, release_version, download_url, package_extension, package_binary_path):
     try:
+        if not is_devtc_installed():
+            print("Please install devtc first!")
+            return
+
         if is_tool_installed(name):
             print("Last version of {} {} is already installed".format(name, release_version))
             return
@@ -77,7 +106,7 @@ def install_tool(name, release_version, url, package_extension, package_binary_p
         os.mkdir(tool_home_path)
         os.mkdir(tool_last_version_path)
 
-        download(url, tool_last_version_path, package_extension)
+        download(download_url, tool_last_version_path, package_extension)
 
         tool_sym_link_path = "{}/default".format(tool_home_path)
         os.symlink(tool_last_version_path, tool_sym_link_path)
@@ -93,13 +122,32 @@ def install_tool(name, release_version, url, package_extension, package_binary_p
 
 
 def remove_tool(name):
+
+    if not is_tool_installed(name):
+        print("{} is not installed!".format(name))
+        return
+
+    print("Removing {}...".format(name))
+
     tool_home_path = "{}/{}".format(DEVTC_HOME_PATH, name)
     shutil.rmtree(tool_home_path, ignore_errors=True)
     tool_link_path = "{}/{}".format(DEVTC_BIN_PATH, name)
     if os.path.islink(tool_link_path):
         os.unlink(tool_link_path)
 
-    print("{} removed successfully".format(name))
+    print("{} successfully removed".format(name))
+
+
+def process_tool(name, process):
+    if name not in ["java", "terraform", "vault"]:
+        print("Unknown tool <{}>!".format(name))
+        return
+    if process not in ["install", "remove"]:
+        print("Did you mean install?")
+        usage()
+        return
+
+    install_tool(name, f_release_version(name), f_download_url(name), f_package_extension(name), f_package_binary_path(name)) if process == "install" else remove_tool(name)
 
 
 def download(url, last_version_path, package_extension):
@@ -153,33 +201,68 @@ def is_tool_installed(name):
     return os.path.isdir(home_path)
 
 
-install_devtc()
-# remove_devtc()
+def usage():
+    print("Usage: devtcApp.py [global options] [tool_to_install]\n" +
+          "\n" +
+          "The available commands are listed below.\n" +
+          "Before installing any tool, you have to install devtcApp first.\n" +
+          "\n" +
+          "Examples:\n" +
+          "devtcApp.py install              Installs devtcApp\n" +
+          "devtcApp.py remove               Removes devtcApp with all installed tools\n" +
+          "devtcApp.py install terraform    Installs terraform tool\n" +
+          "devtcApp.py remove terraform     Removes terraform tool\n" +
+          "\n" +
+          "Main commands:\n" +
+          "install     Prepare your working directory for other commands\n" +
+          "remove      Check whether the configuration is valid"
+          )
 
-java = "java"
-java_release_version = "15"
-java_package_extension = "tar.gz"
-java_package_binary_path = "amazon-corretto-15.0.1.9.1-linux-x64/bin/java"
-java_package_url = "https://corretto.aws/downloads/latest/amazon-corretto-15-x64-linux-jdk.tar.gz"
 
-install_tool(java, java_release_version, java_package_url, java_package_extension, java_package_binary_path)
-# remove_tool(java)
-
-terraform = "terraform"
-terraform_release_version = "0.14.3"
-terraform_package_extension = "zip"
-terraform_package_binary_path = "terraform"
-terraform_package_url = "https://releases.hashicorp.com/terraform/0.14.3/terraform_0.14.3_linux_amd64.zip"
+def f_release_version(name):
+    if name == "java":
+        return "15"
+    elif name == "terraform":
+        return "0.14.3"
+    else:
+        return "1.6.1"
 
 
-install_tool(terraform, terraform_release_version, terraform_package_url, terraform_package_extension, terraform_package_binary_path)
-# remove_tool(terraform)
+def f_download_url(name):
+    if name == "java":
+        return "https://corretto.aws/downloads/latest/amazon-corretto-15-x64-linux-jdk.tar.gz"
+    elif name == "terraform":
+        return "https://releases.hashicorp.com/terraform/0.14.3/terraform_0.14.3_linux_amd64.zip"
+    else:
+        return "https://releases.hashicorp.com/vault/1.6.1/vault_1.6.1_linux_amd64.zip"
 
-vault = "vault"
-vault_release_version = "1.6.1"
-vault_package_extension = "zip"
-vault_package_binary_path = "vault"
-vault_package_url = "https://releases.hashicorp.com/vault/1.6.1/vault_1.6.1_linux_amd64.zip"
 
-install_tool(vault, vault_release_version, vault_package_url, vault_package_extension, vault_package_binary_path)
-# remove_tool(vault)
+def f_package_extension(name):
+    if name == "java":
+        return "tar.gz"
+    elif name == "terraform":
+        return "zip"
+    else:
+        return "zip"
+
+
+def f_package_binary_path(name):
+    if name == "java":
+        return "amazon-corretto-15.0.1.9.1-linux-x64/bin/java"
+    elif name == "terraform":
+        return "terraform"
+    else:
+        return "vault"
+
+
+commands = sys.argv
+number_of_commands = len(commands)
+if number_of_commands == 1:
+    usage()
+elif number_of_commands == 2:
+    process_devtc(commands[1])
+elif number_of_commands == 3:
+    process_tool(commands[1], commands[2])
+else:
+    print("Too many arguments!")
+    usage()
